pyproject.toml
[tool.poetry]
name = "django-configuration-management"
version = "0.1.0"
description = "A management system for Django configuration that encrypts your secrets so you can commit them to the repo."
license = "MIT"
readme = "README.md"
repository = "https://github.com/brno32/..."
authors = ["Alex Drozd <drozdster@gmail.com>", "Tim Sabat <tim@visinote.com>"]
keywords = ["django", "config", "secrets", "settings"]
packages = [
    { include = "django_configuration_management" }
]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.15.0"
pyyaml = "^5.3.1"
cryptography = "^3.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
click = "^7.1.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
upsert_secret = 'django_configuration_management.cli:upsert_secret'
reveal_secrets = 'django_configuration_management.cli:reveal_secrets'
generate_key = 'django_configuration_management.cli:generate_key'
